pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seo-internal-linking"
version = "0.1.0"
description = "CLI tool for automated SEO internal linking using embeddings and Claude API"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "anthropic>=0.40",
    "requests>=2.31",
    "aiohttp>=3.9",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "python-docx>=1.1",
    "openpyxl>=3.1",
    "sentence-transformers>=3.0",
    "numpy>=1.26",
    "tenacity>=8.2",
]

[project.scripts]
seo-linker = "seo_linker.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
